requirements.txt
torch
torchvision
diffusers
transformers
accelerate
opencv-python
Pillow
numpy
matplotlib
tqdm